+++ lib/src/keyframe.dart
@@ -3,6 +3,8 @@ import 'package:boomsheets/src/labeled_point.dart';
 import 'package:boomsheets/src/frametime.dart';
 import 'package:yes_parser/yes_parser.dart';
 
+typedef KeyframeRect = ({Point<int> pos, Point<int> size});
+
 /// A [Keyframe] represents a [rect] on a spritesheet.
 ///
 /// The [rect] can also have a relative [origin] which should
@@ -12,8 +14,8 @@ import 'package:yes_parser/yes_parser.dart';
 /// [isEmpty] represents an invisible [Keyframe] with no [rect] data.
 /// [duration] is how many frames the [rect] should last in [Frametime].
 ///
-/// [computeRect] is the [rect] after [flipX] and [flipY] transformations.
-/// [computeOrigin] is the [origin] after [flipX] and [flipY] transformations.
+/// [flippedRect] is the [rect] after [flipX] and [flipY] transformations.
+/// [flippedRect] is the [origin] after [flipX] and [flipY] transformations.
 ///
 /// Keyframes support scene-graph based renderers with child nodes.
 /// Therefore, arbitrary [points] with a name stored in [LabeledPoint.label]
@@ -22,10 +24,11 @@ import 'package:yes_parser/yes_parser.dart';
 class Keyframe {
   static const Point<int> _pointZeroInt = Point(0, 0);
   static const Point<double> _pointZeroDouble = Point(0.0, 0.0);
-  static const Rectangle<int> _rectZero = Rectangle(0, 0, 0, 0);
+  static const KeyframeRect _rectZero =
+      (pos: _pointZeroInt, size: _pointZeroInt);
 
   List<Attribute> attrs = [];
-  Rectangle<int> rect;
+  KeyframeRect rect;
   Point<int> origin;
   Map<String, LabeledPoint> points = {};
   bool flipX;
@@ -33,13 +36,13 @@ class Keyframe {
   Frametime duration;
   final bool isEmpty;
 
-  Rectangle<int> get computeRect {
+  KeyframeRect get flippedRect {
     if (isEmpty) return _rectZero;
 
-    int x = rect.left;
-    int y = rect.top;
-    int w = rect.width;
-    int h = rect.height;
+    int x = rect.pos.x;
+    int y = rect.pos.y;
+    int w = rect.size.x;
+    int h = rect.size.y;
 
     if (flipX) {
       x = x + w;
@@ -51,44 +54,50 @@ class Keyframe {
       h = -h;
     }
 
-    return Rectangle(x, y, w, h);
+    return (pos: Point<int>(x, y), size: Point<int>(w, h));
   }
 
-  Point<int> get computeOrigin {
+  Point<int> get flippedOrigin {
     if (isEmpty) return _pointZeroInt;
 
     int x = origin.x;
     int y = origin.y;
 
     if (flipX) {
-      x = rect.width - x;
+      x = rect.size.x - x;
     }
 
     if (flipY) {
-      y = rect.height - y;
+      y = rect.size.y - y;
     }
 
     return Point<int>(x, y);
   }
 
-  /// Transforms [computeOrigin] width and height to canonical values [.0-1.0].
+  /// If [considerFlip] is true, this method uses [flippedOrigin]
+  /// and [flippedRect] for calculating the canonical origin values.
+  /// Default value is false.
   ///
-  /// Note that origins can be anywhere outside the [computeRect] and may
-  /// result in negative or even large canonical values.
-  Point<double> get canonicalOrigin {
+  /// This methods transforms [origin] or [flippedOrigin] to
+  /// canonical values (0.0,1.0) inclusive.
+  ///
+  /// Note that origins can be anywhere outside the [rect] or
+  /// [flippedRect] and may result in negative or even large
+  /// canonical values.
+  Point<double> canonicalOrigin({bool considerFlip = false}) {
     if (isEmpty) return _pointZeroDouble;
 
-    final corigin = computeOrigin;
-    final crect = computeRect;
+    final corigin = considerFlip ? flippedOrigin : origin;
+    final crect = considerFlip ? flippedRect : rect;
     double w = 0;
     double h = 0;
 
-    if (crect.width != 0) {
-      w = corigin.x / crect.width;
+    if (crect.size.x != 0) {
+      w = corigin.x / crect.size.x;
     }
 
-    if (crect.height != 0) {
-      h = corigin.y / crect.height;
+    if (crect.size.y != 0) {
+      h = corigin.y / crect.size.y;
     }
 
     return Point<double>(w, h);
@@ -106,7 +115,7 @@ class Keyframe {
   /// Construct an empty, invisible, shapeless keyframe with a [duration].
   Keyframe.empty({required this.duration})
       : isEmpty = true,
-        rect = Rectangle.fromPoints(_pointZeroInt, _pointZeroInt),
+        rect = (pos: _pointZeroInt, size: _pointZeroInt),
         origin = _pointZeroInt,
         flipX = false,
         flipY = false;